Assistant: Core chat support for multiple mapped edits providers (#9411)
Changes logic in core to make a best effort to use the same
extension contributing the current chat provider to apply the edits. So,
if a user is using a Copilot model, Copilot will apply the edits with
the above benefits. Conversely, Assistant models will use the Assistant
implementation.

Addresses #8253