Add keyboard accessibility for data explorer filter popup (#9458)
Addresses #8934
Addresses #7585 

This PR enhances the keyboard accessibility of the Data Explorer's row
filter modal by allowing users to:
1. use arrow keys to navigate between filter elements
2. apply the filters by pressing `Enter` when focused on filter
parameter input fields

The supported filter scenarios for pressing `Enter` are:
* Single parameter filters (e.g., "is equal to", "contains", "starts
with")
* Two parameter filters (e.g., "is between", "is not between") - Press
`Enter` in either input field is supported

For filters without a parameter input field (e.g., "is missing", "is not
missing") the `Enter` key is not supported to allow submitting the
filter since pressing `Enter` is used to open the dropdown.



https://github.com/user-attachments/assets/414a4ed3-c84c-48c6-9efb-da92ad606c7c


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Add keyboard accessibility for data explorer filter popup (#8934,
#7585)

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

@:data-explorer @:web @:win