Turn on data explorer summary panel feature flag (#9512)
Addresses #8527 

This PR exposes the summary panel feature flag as a configuration in the
settings UI and turns on the setting by default.

<img width="884" height="680" alt="Screenshot 2025-09-22 at 1 10 22 PM"
src="https://github.com/user-attachments/assets/36b2da22-ce26-4b7b-a695-cef6d86fd298"
/>

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Enable new data explorer summary panel features (#8527)

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
@:data-explorer @:web @:win

---------

Signed-off-by: Dhruvi Sompura <dhruvi.sompura@posit.co>
Co-authored-by: Julia Silge <julia.silge@gmail.com>