Add keyboard support for context menus in data explorer (#9457)
Addresses #3377 and #9315

This PR adds keyboard accessibility for context menus in the Data
Explorer's main data grid. Users can now access column header, row
header, and cell context menus using keyboard shortcuts.

Since the data grid doesn't support focusing/selecting the column
headers and row headers via keyboard navigation, the cursor remains on
the data cells. Different keyboard shortcuts will allow the user to open
the column header or row header context menu of the row or column that
they are currently on.

`Shift + F10` - Opens the cell context menu for the cell that has focus
`Ctrl + Shift + F10` - Opens the column header context menu based off
the cell that has focus
`Alt + Shift + F10` - Opens the row header context menu based off the
cell that has focus

When the context menu for a row header or column header is selected, the
entire row or column will be selected/focused as well. This change
allows copying an entire row or column to work as expected (#9315).

**Copy Column or Row**


https://github.com/user-attachments/assets/b3436d03-0449-457d-a3ee-b9385e71d6fe


**Keyboard Shortcuts for Context Menu**



https://github.com/user-attachments/assets/f36acc24-4c05-4fe9-81ae-bcc13fbf04a9

--

Update from 09/22/2025 to resolve the keyboard shortcut issue mentioned
in
https://github.com/posit-dev/positron/pull/9457#pullrequestreview-3247604236:

- we now have separate actions for each data explorer context menu
(column, row, cell) which allows us to show the commands in the command
palette, and also exposes the keyboard shortcuts to users via settings
- fixed a bug where the data explorer copy command was not showing up in
the command palette menu because of the precondition that required the
data explorer to have focus (which was always false because focus would
be lost once the command palette was opened)

<img width="1468" height="544" alt="Screenshot 2025-09-22 at 5 05 48 PM"
src="https://github.com/user-attachments/assets/b34117d7-56fa-4b8f-a0dd-63fe6d759a24"
/>


https://github.com/user-attachments/assets/f92cea39-dc8b-4586-ba7d-8d71cc320a68

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Add keyboard support for context menus in data explorer (#3377)

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
@:data-explorer @:web @:win