cleanup unused data explorer context key (#9540)
This is a follow up PR to
https://github.com/posit-dev/positron/pull/9457 that contains some code
cleanup.

In https://github.com/posit-dev/positron/pull/9457 we removed all use
cases for `PositronDataExplorerFocused` context key (checks if data
explorer tab is focused) and won't need it in the future since we have
the `POSITRON_DATA_EXPLORER_IS_ACTIVE_EDITOR` context key (checks if the
active editor tab is a data explorer instance).

This PR removes the context key and all the code paths used to set the
context key value.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes
The main thing to verify here, is that general focus behavior for the
data explorer editor tab is still working. That behavior shouldn't be
effected by this change, since the context key was only used to show the
data grid context menus.

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

@:data-explorer