Replace 'request' with 'axios' for more reliable supervisor connectivity (#9421)
This change is intended to holistically resolve a number of issues
around HTTP proxies getting in the way of communication between
Kallichore and Positron. Many of the issues had to do with the
`HTTP_PROXY` and `NO_PROXY` environment variables, which are read by the
`request` npm package.

The fix is to remove the `request` library _entirely_, which we need to
do soon anyway, since this library has been [deprecated for over 5
years](https://github.com/request/request/issues/3142).

Unfortunately, removing the dependency is not straightforward because we
don't depend on it directly; it is the core library used by the OpenAPI
`typescript-node` code generator, and it is therefore the generated API
interface, not Positron itself, that uses the library. There is a lot of
support for [getting rid of `request` there,
too](https://github.com/OpenAPITools/openapi-generator/issues/5334), but
little chance of it happening soon.

In order to get rid of `request`, then, we need to use an entirely
different code generator. I evaluated `typescript-fetch` and
`typescript-axios` and selected
[typescript-axios](https://openapi-generator.tech/docs/generators/typescript-axios/)
because (a) Axios is an actively maintained project, and (b) it works
out of the box on the node.js runtime, whereas `typescript-fetch` only
generates browser-compatible code that would require heavy
hand-modification every time we needed to rev the API.

Axios has a `proxy: false` option that just ignores all the
proxy-related environment variables, which is what we want here; not
only does this prevent them from getting in the way, it also allows us
to preserve their values for use in subprocesses (which may need them!).

A few bonus changes while I was in the code:

- started improving test coverage; there are now tests for all 3
transport types
- drive-by fix for issue on Windows that shows up when you try to turn
on the idle shutdown timeout

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fix errors starting Python and R sessions on some systems when
`HTTP_PROXY` is set (#9215)
- Fix errors starting Python and R sessions on Windows when the Kernel
Supervisor's _Shutdown Timeout_ setting is used (#6368)



### QA Notes

The largest risk of regressions here are around (a) error handling,
which Axios does differently than request, and (b) named pipes on
Windows, which also work somewhat differently in the new world.

It's also worth checking both native IPC and TCP transports on at least
two platforms. Note that this change does not go into effect until a new
supervisor starts, so after you change the transport you'll need to
fully restart Positron to pick it up.

I did run a full test suite on this change (after the Axios refactor,
not on the latest bits) due to its scope.
https://github.com/posit-dev/positron/actions/runs/17684238720