support previewing ibis tables in data explorer (#9393)
Support previewing `ibis` tables in the data explorer. This takes the
same route as data viewed from the Connections pane, as in, it
implements a 1000 row preview for the data. Because of this, we can
largely rely on the pandas backend since anything that uses
`table.execute()` returns a pandas dataframe.

Follow up work is outlined in a few issues:
- For having a "real" but lazy summary panel:
https://github.com/posit-dev/positron/issues/9516
- For having convert to code features:
https://github.com/posit-dev/positron/issues/9514

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #6515 Support to preview `ibis` tables in the Connections pane.

#### Bug Fixes

- N/A


### QA Notes

```python
import ibis

data = [
    {"id": 1, "name": "Alice", "age": 30, "score": 85.5},
    {"id": 2, "name": "Bob",   "age": 25, "score": 92.3},
    {"id": 3, "name": "Carol", "age": 27, "score": 70.1},
    {"id": 4, "name": "Dave",  "age": 35, "score": 60.2},
]

t = ibis.memtable(data)
%view t
```

---------

Signed-off-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>