Assistant: Improve `executeCode` context and likelihood of correct usage (#9557)
This change contains a couple of tweaks to `executeCode` to address some
problems we've seen with the way the model uses it.

Most importantly, the model must now tell Assistant _which session_ to
run the code in rather than only specifying a language. This makes it
(almost) impossible for the model to try to run, say, bash code in an R
session.

(To do this it was necessary to plumb the session ID all the way through
the API, which I think will be helpful in other contexts; please excuse
the parameter ordering issues as I didn't want to shuffle parameters
that extensions are already using)

Secondly, we now tell the model to prefer this tool over other ways of
running code when a session is available. It likely isn't foolproof but
empirically it helps Copilot Chat realize it can just run stuff right in
the Positron console instead of trying to start a terminal.

<img width="451" height="611" alt="image"
src="https://github.com/user-attachments/assets/e6548e68-ab16-4c48-a2a1-1b136ba0c7d7"
/>


Addresses https://github.com/posit-dev/positron/issues/8727
Addresses https://github.com/posit-dev/positron/issues/9511

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix issue causing Assistant to occasionally try to execute code with
the wrong language in a Positron console (#8727)


### QA Notes

To help Copilot Chat realize it can use this tool, I had to tone down
the language around only using it when necessary. I think that language
was partially responsible for Copilot always reaching for the Terminal
(now here we have two ways of running code but only _one_ of them says
not to use it unless I have to...)

So the main risk of this change is that it will reintroduce some of
Claude Sonnet 4's tendency to want to run code ALL THE TIME even if it
doesn't need to. Please let me know if it feels like it's doing this.

Tags: `@:assistant`