Assistant: Switch to using `languageModelSystem` API for system prompts (#9587)
In this PR we remove the use of the `modelOptions.system` mechanism for
setting system prompts, and instead switch to using the
`languageModelSystem` API (which involves sending messages with the
`System` role) as used by Copilot Chat.

With these changes, the mechanism is properly supported for both the
Anthropic and Vercel AI SDK:
- For Anthropic we convert System messages into a `system` property to
be passed to the API.
- For Vercel we convert the System messages into a single system message
prepended to the messages array.

The option to provide a `modelOptions.system` parameter remains for now,
but we can remove it once extensions such as Databot have also migrated
to use the newer mechanism.

This is a follow on to #9553, and cherry-picks a commit from there. So
it should be merged before this PR.

Addresses #9434, by enabling Databot to send its system prompt to both
Assistant and Copilot models without special casing once it has migrated
to the new `languageModelSystem` API.

### QA Notes

This is an internal implementation detail, so there should be no user
facing changes.

Chatting to both Assistant and Copilot registered models should be as
normal, including with the Positron system prompt attached. This can be
confirmed by watching the Assistant and GitHub Copilot Chat output
panels in trace mode. For Copilot registered models, the Copilot and
Assistant system prompts should be concatenated.

<img width="1607" height="224" alt="Screenshot 2025-09-25 at 16 56 35"
src="https://github.com/user-attachments/assets/8e2053d6-0906-4c5e-bf1b-a3041008922c"
/>