Add notebook consoles; fix reload (#9550)
This change introduces 'notebook consoles', which are consoles that can
be used as a sort of scratchpad for notebooks. A notebook console is
attached to a notebook session; code you run in the console affects the
notebook session, but doesn't appear in the notebook itself, making it
ideal for one-off commands or computations better suited for a REPL than
a throwaway cell.

Notebook consoles appear in the Console tab with a notebook icon instead
of a language logo, and the filename of the notebook listed beside them.

<img width="283" height="161" alt="image"
src="https://github.com/user-attachments/assets/0f4ea9da-2bc5-4f0b-8539-e6b13c833d76"
/>

While useful, these consoles can be kind of confusing to new users
because it gives you two places to run code, and some things work a
little differently when the kernel is in notebook (vs. REPL) mode.
Consequently, the notebook console does not appear by default.

The new command _Console: Show Notebook Console_ will trigger a notebook
console; there's also a button on the new notebooks (and the old ones,
not pictured):

<img width="367" height="69" alt="image"
src="https://github.com/user-attachments/assets/5a95779a-c158-45af-9a3e-9bff0cc85b9c"
/>

For users who always want one of these available, there's an option to
auto-start one for every notebook:

<img width="401" height="100" alt="image"
src="https://github.com/user-attachments/assets/8564b4a7-6c2d-4ec5-8a01-10305e39710c"
/>

While I was in the code, I noticed that reload for notebook sessions has
been broken for a while (see #9453). The problem was a fairly
complicated start loop wherein restoring the session on reload triggered
a pre-registration of the associated runtime, which triggered the
associated notebook kernel to get created, which triggered a session
start again. The fix there is to avoid trying to track the mapping
between sessions and notebook URIs in the runtime kernel service, and to
wait for sessions to be reconnected before trying to auto-start them for
notebooks.


Addresses https://github.com/posit-dev/positron/issues/3801.
Addresses https://github.com/posit-dev/positron/issues/9453.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Add an option to attach a console to your notebook's R or Python
session for one-off commands and interactive work (#3801)

#### Bug Fixes

- N/A


### QA Notes

After you create the notebook console, any code you execute in the
notebook will also appear in the notebook's console. This is normal;
it's meant to help you see what's been run leading up to your current
state.

In notebook consoles, running commands that generate non-textual output
(like graphs or HTML content) will often not work correctly, because in
notebook mode this content typically goes to a cell.



This change adds an E2E test to ensure that regression w/ reloading the
window does not occur again.