Data Explorer: Always treat first row of CSV as header row (#9607)
Addresses #8860. We won't currently have options to properly distinguish
between data files that have a header row and those that don't, and
DuckDB's inference logic to determine the presence of a header can fail
in certain files (setting "no header" for files that do have a header).
In our discussion we decided that refusing to guess/infer was a better
approach, and since the "with header" option is the most common one, it
should be the default behavior. We will eventually have to figure out
how to allow users to look at headerless CSV files (probably with some
option dropdown menu, not sure).