notebooks: prompt user to change session working directory if saving untitled nb (#9360)
Addresses https://github.com/posit-dev/positron/issues/9022. Adds a
modal that pops up when a notebook is moved (including saving an
untitled notebook) that allows the session working directory to be
changed without destroying the session.

This doesn't currently work for "moving" a notebook (dragging it through
the file explorer or saving-as) because of
https://github.com/posit-dev/positron/issues/9089. For some reason that
workflow isn't triggering the `saveAs` code like it should be, which
would also correctly move the URI. I'll scope that fix to that issue.

<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- https://github.com/posit-dev/positron/issues/9022 When saving untitled
notebooks, you can now fix their working directory without losing data.

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->
@:notebooks @:web @:win

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
A bunch of cases I can think of. In each case there shouldn't be data
loss (any existing variables should still be there).

- Save an untitled notebook outside of a workspace. The modal should pop
up if you save it somewhere besides your home dir.
- Save an untitled notebook inside of a workspace. The modal should pop
up if you save it somewhere besides the workspace root.
- Try without/with the `notebook.workingDirectory` setting set. If set,
the modal should only pop up if the resolved setting would change.
- Watch out for https://github.com/posit-dev/positron/issues/9631. In
fact that issue may make it impossible to validate this setting
behavior.
- Try clicking on "keep" vs "update" and ensure they behave as expected
- Try web and windows