Add support for 0-row dataframes in Data Explorer (#9628)
Addresses #9619

This PR adds support for zero-row datasets in the main data grid. The
data grid would only fetch and cache data if the number of rows AND
columns in the dataset was not zero. When a dataset with zero rows was
opened, the column header title would be missing because we skipped the
entire data fetching step.

We no longer block data fetching if there are no rows in the dataframe.
The caching layer has been updated so we can fetch the column schema if
there are columns present.