Remove experimental data explorer config (#9604)
Addresses #8528
Related PR for Ark: https://github.com/posit-dev/ark/pull/931

Addresses the second half of #8528 which was to remove the
`dataExplorer.Experimental` config setting since it is confusing for
users when they see it.

For features that require frontend and backend coordination for the data
explorer, we have a system for declaring what features are supported.
The `SupportedFeatures` interface communicates the list of features that
are supported/unsupported by each backend. Each feature declares its
support status via a value from the `SupportStatus` enum. This enum has
three possible options: `Supported`, `Unsupported`, and `Experimental`.

We are removing the `Experimental` option and
`dataExplorer.Experimental` config setting since it isn't really used
for development (I think?). We recently added a new feature to the
`SupportedFeatures` list, convert to code, and that feature didn't
utilize the `Experimental` setting.

This change involved updating the data explorer backend rpc comm json to
remove the `Experimental` value.

This is my first time updating comms, so extra eyes on the changes would
be nice. I was able to test the changes agains a local build of ark but
additional testing would be appreciated to make sure I didn't miss
anything.

To get this change across the finish line, I believe we want to merge
the Ark PR first, release Ark (?), and then pick up the new Ark version
in this PR before this PR is merged.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Remove experimental Data Explorer config setting (#8528)

#### Bug Fixes

- N/A


### QA Notes

There isn't anything specific to test here since this setting was never
used. We'll just want to verify that `dataExplorer.Experimental` no
longer shows up in Settings:

<img width="1273" height="417" alt="Screenshot 2025-09-25 at 3 40 17 PM"
src="https://github.com/user-attachments/assets/73c18519-1d2f-4a8e-997c-7b9db2758053"
/>

@:data-explorer @:web @:win